+++ packages/x6/src/geometry/line.ts
@@ -1,9 +1,9 @@
 import { Path } from './path'
 import { Point } from './point'
 import { Ellipse } from './ellipse'
+import { Geometry } from './geometry'
 import { Polyline } from './polyline'
 import { Rectangle } from './rectangle'
-import { Geometry } from './geometry'
 
 export class Line extends Geometry {
   public start: Point
@@ -41,6 +41,9 @@ export class Line extends Geometry {
     return this.center
   }
 
+  /**
+   * Rounds the line to the given `precision`.
+   */
   round(precision: number = 0) {
     this.start.round(precision)
     this.end.round(precision)
@@ -61,28 +64,46 @@ export class Line extends Geometry {
     return this
   }
 
+  /**
+   * Rotate the line by `angle` around `origin`.
+   */
   rotate(angle: number, origin?: Point.PointLike | Point.PointData) {
     this.start.rotate(angle, origin)
     this.end.rotate(angle, origin)
     return this
   }
 
+  /**
+   * Scale the line by `sx` and `sy` about the given `origin`. If origin is not
+   * specified, the line is scaled around `0,0`.
+   */
   scale(sx: number, sy: number, origin?: Point.PointLike | Point.PointData) {
     this.start.scale(sx, sy, origin)
     this.end.scale(sx, sy, origin)
     return this
   }
 
+  /**
+   * Returns the length of the line.
+   */
   length() {
     return Math.sqrt(this.squaredLength())
   }
 
+  /**
+   * Useful for distance comparisons in which real length is not necessary
+   * (saves one `Math.sqrt()` operation).
+   */
   squaredLength() {
     const dx = this.start.x - this.end.x
     const dy = this.start.y - this.end.y
     return dx * dx + dy * dy
   }
 
+  /**
+   * Scale the line so that it has the requested length. The start point of
+   * the line is preserved.
+   */
   setLength(length: number) {
     const total = this.length()
     if (!total) {
@@ -94,8 +115,7 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns the vector (point) of the line with length equal to
-   * length of the line.
+   * Returns the vector of the line with length equal to length of the line.
    */
   vector() {
     return new Point(this.end.x - this.start.x, this.end.y - this.start.y)
@@ -104,8 +124,11 @@ export class Line extends Geometry {
   /**
    * Returns the angle of incline of the line.
    *
-   * The function returns `NaN` if the two endpoints of the line
-   * both lie at the same coordinates.
+   * The function returns `NaN` if the start and end endpoints of the line
+   * both lie at the same coordinates(it is impossible to determine the angle
+   * of incline of a line that appears to be a point). The
+   * `line.isDifferentiable()` function may be used in advance to determine
+   * whether the angle of incline can be computed for a given line.
    */
   angle() {
     const horizontal = new Point(this.start.x + 1, this.start.y)
@@ -137,32 +160,30 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns the point on the line that lies closest to point.
+   * Returns the point on the line that lies closest to point `p`.
    */
   closestPoint(p: Point.PointLike | Point.PointData) {
     return this.pointAt(this.closestPointNormalizedLength(p))
   }
 
   /**
-   * Returns the length of the line up to the point that lies
-   * closest to point.
+   * Returns the length of the line up to the point that lies closest to point `p`.
    */
   closestPointLength(p: Point.PointLike | Point.PointData) {
     return this.closestPointNormalizedLength(p) * this.length()
   }
 
   /**
-   * Returns a line that is tangent to the line at the point that
-   * lies closest to point.
+   * Returns a line that is tangent to the line at the point that lies closest
+   * to point `p`.
    */
   closestPointTangent(p: Point.PointLike | Point.PointData) {
     return this.tangentAt(this.closestPointNormalizedLength(p))
   }
 
   /**
-   * Returns the normalized length (distance from the start of the
-   * line / total line length) of the line up to the point that lies
-   * closest to point.
+   * Returns the normalized length (distance from the start of the line / total
+   * line length) of the line up to the point that lies closest to point.
    */
   closestPointNormalizedLength(p: Point.PointLike | Point.PointData) {
     const product = this.vector().dot(new Line(this.start, p).vector())
@@ -177,8 +198,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns a point on the line that lies `rate` (normalized length)
-   * away from the beginning of the line.
+   * Returns a point on the line that lies `rate` (normalized length) away from
+   * the beginning of the line.
    */
   pointAt(ratio: number) {
     const start = this.start
@@ -196,8 +217,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns a point on the line that lies length away from the
-   * beginning of the line.
+   * Returns a point on the line that lies length away from the beginning of
+   * the line.
    */
   pointAtLength(length: number) {
     const start = this.start
@@ -220,8 +241,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Divides the line into two lines at the point that lies `rate`
-   * (normalized length) away from the beginning of the line.
+   * Divides the line into two lines at the point that lies `rate` (normalized
+   * length) away from the beginning of the line.
    */
   divideAt(ratio: number) {
     const dividerPoint = this.pointAt(ratio)
@@ -232,8 +253,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Divides the line into two lines at the point that lies length
-   * away from the beginning of the line.
+   * Divides the line into two lines at the point that lies length away from
+   * the beginning of the line.
    */
   divideAtLength(length: number) {
     const dividerPoint = this.pointAtLength(length)
@@ -243,6 +264,9 @@ export class Line extends Geometry {
     ]
   }
 
+  /**
+   * Returns `true` if the point `p` lies on the line. Return `false` otherwise.
+   */
   containsPoint(p: Point.PointLike | Point.PointData) {
     const start = this.start
     const end = this.end
@@ -265,13 +289,17 @@ export class Line extends Geometry {
     return true
   }
 
+  /**
+   * Returns an array of the intersection points of the line with another
+   * geometry shape.
+   */
   intersect(shape: Line | Rectangle | Polyline | Ellipse): Point[] | null
   intersect(shape: Path, options?: Path.Options): Point[] | null
   intersect(
     shape: Line | Rectangle | Polyline | Ellipse | Path,
     options?: Path.Options,
   ): Point[] | null {
-    const ret = shape.intersectionWithLine(this, options)
+    const ret = shape.intersectsWithLine(this, options)
     if (ret) {
       return Array.isArray(ret) ? ret : [ret]
     }
@@ -280,10 +308,10 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns the intersection point of the line with another line.
-   * Returns `null` if no intersection exists.
+   * Returns the intersection point of the line with another line. Returns
+   * `null` if no intersection exists.
    */
-  intersectionWithLine(line: Line) {
+  intersectsWithLine(line: Line) {
     const pt1Dir = new Point(
       this.end.x - this.start.x,
       this.end.y - this.start.y,
@@ -320,15 +348,21 @@ export class Line extends Geometry {
     )
   }
 
+  /**
+   * Returns `true` if a tangent line can be found for the line.
+   *
+   * Tangents cannot be found if both of the line endpoints are coincident
+   * (the line appears to be a point).
+   */
   isDifferentiable() {
     return !this.start.equals(this.end)
   }
 
   /**
-   * Returns the perpendicular distance between the line and point.
-   * The distance is positive if the point lies to the right of the
-   * line, negative if the point lies to the left of the line, and
-   * `0` if the point lies on the line.
+   * Returns the perpendicular distance between the line and point. The
+   * distance is positive if the point lies to the right of the line, negative
+   * if the point lies to the left of the line, and `0` if the point lies on
+   * the line.
    */
   pointOffset(p: Point.PointLike | Point.PointData) {
     const ref = Point.clone(p)
@@ -365,8 +399,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns a line tangent to the line at point that lies `rate`
-   * (normalized length) away from the beginning of the line.
+   * Returns a line tangent to the line at point that lies `rate` (normalized
+   * length) away from the beginning of the line.
    */
   tangentAt(ratio: number) {
     if (!this.isDifferentiable()) {
@@ -384,8 +418,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns a line tangent to the line at point that lies `length`
-   * away from the beginning of the line.
+   * Returns a line tangent to the line at point that lies `length` away from
+   * the beginning of the line.
    */
   tangentAtLength(length: number) {
     if (!this.isDifferentiable()) {
@@ -403,8 +437,8 @@ export class Line extends Geometry {
   }
 
   /**
-   * Returns which direction the line would have to rotate in order to
-   * direct itself at a point.
+   * Returns which direction the line would have to rotate in order to direct
+   * itself at a point.
    *
    * Returns 1 if the given point on the right side of the segment, 0 if its
    * on the segment, and -1 if the point is on the left side of the segment.
@@ -437,6 +471,9 @@ export class Line extends Geometry {
     return ccw < 0.0 ? -1 : ccw > 0.0 ? 1 : 0
   }
 
+  /**
+   * Return `true` if the line equals the other line.
+   */
   equals(l: Line) {
     return (
       l != null &&
@@ -447,6 +484,9 @@ export class Line extends Geometry {
     )
   }
 
+  /**
+   * Returns another line which is a clone of the line.
+   */
   clone() {
     return new Line(this.start, this.end)
   }